requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
web3==6.20.1
eth-account==0.10.0
SQLAlchemy==2.0.32
